faucet/requirements.txt
aiohttp==3.7.4.post0
asn1crypto==1.4.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
base58==2.1.1
cached-property==1.5.2
certifi==2019.6.16
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
coincurve==16.0.0
cytoolz==0.11.2
discord.py==1.7.3
ecdsa==0.17.0
eth-hash==0.3.2
eth-keys==0.4.0
eth-typing==2.2.2
eth-utils==1.10.0
ethereum==2.3.2
future==0.18.2
idna==2.8
idna-ssl==1.1.0
more-itertools==8.12.0
multidict==5.2.0
mypy-extensions==0.4.3
pbkdf2==1.3
py-bip39-bindings==0.1.8
py-ecc==5.2.0
py-ed25519-bindings==1.0.1
py-sr25519-bindings==0.1.4
pycparser==2.21
pycryptodome==3.12.0
pyethash==0.1.27
pysha3==1.0.2
PyYAML==6.0
repoze.lru==0.7
requests==2.27.1
rlp==1.2.0
scalecodec==1.0.28
scrypt==0.8.19
six==1.16.0
substrate-interface==1.1.7
substrateinterface==0.9.3
toolz==0.11.2
typing_extensions==4.0.1
urllib3==1.25.3
websocket-client==1.2.3
websockets==8.1
xxhash==1.3.0
yarl==1.7.2
setuptools==45.2.0